cs/devtools/requirements.txt
black==24.10.0
click==8.1.7
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.3.6
typing_extensions==4.12.2
packaging==24.1
pyyaml==6.0.2